cbplus_test/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.1
django-extended-choices==1.3.3
django-filter==2.3.0
django-model-utils==3.1.2
djangorestframework==3.11.1
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
Markdown==3.2.2
MarkupSafe==1.1.1
psycopg2-binary==2.8.5
Pygments==2.6.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
